refactor(chat): extract hot_questions.py from the chat god file
Verbatim move of DEFAULT_HOT_QUESTIONS / HOT_QUESTION_CATEGORIES /
get_hot_questions / get_random_hot_questions / get_hot_question_prompt.
Symbols re-imported into app.services.chat so the api layer's imports
and tests' patch targets are unchanged. 819 tests pass.
